refactor(agents): move claude and codex into adapter folders
