+++ packages/browseros-agent/apps/server/src/lib/agents/claude/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @license
+ * Copyright 2025 BrowserOS
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+export {
+  ClaudeRuntime,
+  type ClaudeRuntimeConfig,
+  type ConfigureClaudeRuntimeOptions,
+  configureClaudeRuntime,
+  getClaudeRuntime,
+  prepareClaudeCodeContext,
+} from './runtime'
