+++ packages/browseros-agent/apps/server/src/lib/agents/claude/runtime.ts
@@ -15,9 +15,9 @@ import {
   prepareBrowserosManagedContext,
 } from '../acpx-agent-common'
 import { resolveAgentRuntimePaths } from '../acpx-runtime-context'
-import { HostProcessAgentRuntime } from './host-process-agent-runtime'
-import { getAgentRuntimeRegistry } from './registry'
-import type { RuntimeDescriptor } from './types'
+import { HostProcessAgentRuntime } from '../runtime/host-process-agent-runtime'
+import { getAgentRuntimeRegistry } from '../runtime/registry'
+import type { RuntimeDescriptor } from '../runtime/types'
 
 const CLAUDE_BINARY = 'claude'
