+++ packages/browseros-agent/apps/server/src/lib/agents/codex/index.ts
@@ -1,0 +1,14 @@
+/**
+ * @license
+ * Copyright 2025 BrowserOS
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+export {
+  CodexRuntime,
+  type CodexRuntimeConfig,
+  type ConfigureCodexRuntimeOptions,
+  configureCodexRuntime,
+  getCodexRuntime,
+  prepareCodexContext,
+} from './runtime'
