+++ packages/browseros-agent/apps/server/src/lib/agents/runtime/index.ts
@@ -5,22 +5,6 @@
  */
 
 export type { AgentRuntime } from './agent-runtime'
-export {
-  ClaudeRuntime,
-  type ClaudeRuntimeConfig,
-  type ConfigureClaudeRuntimeOptions,
-  configureClaudeRuntime,
-  getClaudeRuntime,
-  prepareClaudeCodeContext,
-} from './claude-host-process-runtime'
-export {
-  CodexRuntime,
-  type CodexRuntimeConfig,
-  type ConfigureCodexRuntimeOptions,
-  configureCodexRuntime,
-  getCodexRuntime,
-  prepareCodexContext,
-} from './codex-host-process-runtime'
 export { ContainerAgentRuntime } from './container-agent-runtime'
 export { ActionNotSupportedError, RuntimeNotReadyError } from './errors'
 export {
